% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.TVG.R
\name{apply.TVG}
\alias{apply.TVG}
\title{Adds or removes TVG.}
\usage{
apply.TVG(x, beams, rm = FALSE, linear = TRUE, TVG.exp = 2, Ro = NULL,
  thr1m = FALSE)
}
\arguments{
\item{x}{is the input acoustic data arranged in an array having the radial part along the first dimension, the beams along the second or second and third, and time steps and other optional dimensions along the last dimension (lenb,numb,numt).}

\item{beams}{is the list of beams inputs including "lenb" (lengths of beams), "sint" (pulse length), "absr" (absorption coefficent of the beams) and "asps" (average speed of sound).}

\item{rm}{is TRUE to remove the TVG instead of adding it.}

\item{linear}{is TRUE if x is in linear volume backscatter values (sv) and FALSE if x is in logarithmic volume backscatter values (Sv).}

\item{TVG.exp}{is the exponent of the eamotric spreading of the sound wave, theoretically 2 for Sv and 4 for TS.}
}
\value{

}
\description{
Adds or removes TVG.
}
\examples{
\dontrun{}

}
