% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.TVG_old.R
\name{get.TVG_old}
\alias{get.TVG_old}
\title{Calculate the Time Varied Gain function.}
\usage{
get.TVG_old(x, beams, linear = TRUE, TVG.exp = 2, TVG.Corr = -2)
}
\arguments{
\item{x}{is the input acoustic data arranged in an array having the radial part along the first dimension, the beams along the second or second and third, and time steps and other optional dimensions along the last dimension (lenb,numb,numt).}

\item{beams}{is the list of beams inputs including "lenb" (lengths of beams), "sint" (pulse length), "absr" (absorption coefficent of the beams) and "asps" (average speed of sound).}

\item{linear}{is TRUE if x is in linear volume backscatter values (sv) and FALSE if x is in logarithmic volume backscatter values (Sv).}

\item{TVG.exp}{is the exponent of the eamotric spreading of the sound wave, theoretically 2 for Sv and 4 for TS.}

\item{delta}{is used in Gavins work.}

\item{R0}{is used in Gavins work, shifting the first voxel to be imaginarily placed inside of the sonar.}
}
\value{

}
\description{
Calculate the Time Varied Gain function.
}
\examples{
\dontrun{}

}
