% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDgLen.R
\name{getDgLen}
\alias{getDgLen}
\title{Calcualtes the data gram length of a datagram to write to a Simrad raw file.}
\usage{
getDgLen(thisdata, thisdatagramName, nBytesDgHeader, nBytesSampledataInfo)
}
\arguments{
\item{thisdata}{is a list of the datagram.}

\item{thisdatagramName}{is the type of datagram.}

\item{nBytesDgHeader}{is the number of bytes of the datagram header.}

\item{nBytesSampledataInfo}{is the number of bytes of the sample data information.}
}
\description{
Calcualtes the data gram length of a datagram to write to a Simrad raw file.
}
\examples{
\dontrun{}

}
