% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_GetRAW.R
\name{getMode}
\alias{getMode}
\title{Reads the sample data from a Simrad raw1 file.}
\usage{
getMode(x, ...)
}
\arguments{
\item{x}{is a raw vector read from a Simrad raw file.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\value{

}
\description{
Reads the sample data from a Simrad raw1 file.
}
\examples{
\dontrun{}

}
