% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRangeOffsetInUnitsOfSamples.R
\name{getRangeOffsetInUnitsOfSamples}
\alias{getRangeOffsetInUnitsOfSamples}
\title{Shifts sonar data by an empirical range offset.}
\usage{
getRangeOffsetInUnitsOfSamples(beams)
}
\arguments{
\item{beams}{A list containing the fields rofs (range offset in meters) and one of rres (range resolution) og the pair asps (average speed of sound) and sint (sampling interval).}
}
\value{

}
\description{
Shifts sonar data by an empirical range offset.
}
\examples{
\dontrun{}

}
