% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer.base.b.R
\name{integer.base.b}
\alias{integer.base.b}
\title{Get integer base of the input. Based on code by Lars Nonboe Andersen, Simrad.}
\usage{
integer.base.b(x, b = 2, endian = "little")
}
\arguments{
\item{x}{is the value of sampledata$datatype.}

\item{b}{is the base.}

\item{endian'}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31). NA}
}
\value{

}
\description{
Get integer base of the input. Based on code by Lars Nonboe Andersen, Simrad.
}
\examples{
\dontrun{}

}
