% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMEA2vessel.R
\name{NMEA2vessel}
\alias{NMEA2vessel}
\title{Return vessel times, positions, headings and speeds located in an NMEA atring.}
\usage{
NMEA2vessel(x, cleanNMEA = 1, interpolate = TRUE)
}
\arguments{
\item{x}{is an NMEA string.}

\item{cleanNMEA}{is FALSE to return the data as read, without shaving off incomplete time steps, 1 to remove incomplete and duplicate time steps, and 2 to additionally clean missing info at the end.}
}
\value{

}
\description{
Return vessel times, positions, headings and speeds located in an NMEA atring.
}
\examples{
\dontrun{}

}
