% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsetSamples.R
\name{offsetSamples}
\alias{offsetSamples}
\title{Shifts sonar data by an empirical range offset.}
\usage{
offsetSamples(x, beams)
}
\arguments{
\item{x}{The acoustic data to be shifted by the offset}

\item{rofs}{The empirical range offset.}

\item{rres.exp}{The radial resolution of the sonar (c * dt / 2, where c is speed of sound and dt is sampling interval duration)}
}
\value{

}
\description{
Shifts sonar data by an empirical range offset.
}
\examples{
\dontrun{}

}
