% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readAllNMEA}
\alias{readAllNMEA}
\title{Read only vessel data from a Simrad raw file.}
\usage{
readAllNMEA(f, NMEA.out = FALSE)
}
\arguments{
\item{f}{The path to the raw file.}

\item{NMEA.out}{logiacl: If TRUE retun a list of the NMEA strings and the vessel info.}
}
\description{
Read only vessel data from a Simrad raw file.
}
