% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcalfile.R
\name{readcalfile}
\alias{readcalfile}
\title{Reads calibration files in the xml format used by LSSS.}
\usage{
readcalfile(x)
}
\value{

}
\description{
Reads calibration files in the xml format used by LSSS.
}
\examples{
\dontrun{}

}
