% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_extractBeams.R
\name{readEKRaw_extractBeams}
\alias{readEKRaw_extractBeams}
\title{Extracts one or more beams from raw data read by readEKRaw().}
\usage{
readEKRaw_extractBeams(x, beamnr = NULL, drop = FALSE)
}
\value{

}
\description{
Extracts one or more beams from raw data read by readEKRaw().
}
\examples{
\dontrun{}

}
