% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_extractPings.R
\name{readEKRaw_extractPings}
\alias{readEKRaw_extractPings}
\title{Modify parts of EK80/WBAT .raw files}
\usage{
readEKRaw_extractPings(filename, newDir, t = "all", addMRU = TRUE,
  fixBeamType = TRUE, fixSlope = TRUE, addTransducerSerialNumber = "114",
  endian = "little", msg = TRUE)
}
\arguments{
\item{filename}{The path to directory of raw files or a vector of the paths to the raw files.}

\item{newDir}{The directory in which to put the modified files.}

\item{addMRU}{logical; if TRUE the funciton inserts a dummy MRU0 datagram prior to RAW3 datagrams.}

\item{fixBeamType}{logical; if TRUE the BeamType is fixed to 1.}

\item{fixSlope}{logical; if TRUE the slope is set to a default.}

\item{endian}{the endianness of the file, defaulted to "little".}

\item{msg}{logical: if TRUE print a time bar during file conversion.}

\item{addTransducerSeralNumber}{A string giving the transducer serial number.}
}
\value{
A vector of file names of the modified files.
}
\description{
Provides the facility to modify parts of EK80/WBAT .raw files so that these files can be read by the Large Scale Survey System (LSSS) software.
}
