% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_GetAcosuticData.R
\name{readEKRaw_GetAcosuticData}
\alias{readEKRaw_GetAcosuticData}
\title{Gets acosutic (or electrical) data from a raw vector read from a raw file.}
\usage{
readEKRaw_GetAcosuticData(x, sampledata, pos, dgName = c("RAW0", "RAW1"),
  endian = "little", complex.out = FALSE)
}
\arguments{
\item{x}{is the raw vector read from the raw file.}

\item{sampledata}{is the list read so far using e.g., \code{\link{readEKRaw_GetRAW1}}.}

\item{dgName}{is the datagram name.}
}
\value{

}
\description{
Gets acosutic (or electrical) data from a raw vector read from a raw file.
}
\examples{
\dontrun{}

}
