% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_GetConfigHeader_temp.R
\name{readEKRaw_GetConfigHeader}
\alias{readEKRaw_GetConfigHeader}
\title{Reads the configuration header from a Simrad raw0 file.}
\usage{
readEKRaw_GetConfigHeader(x, endian = "little")
}
\arguments{
\item{x}{is a raw vector read from a Simrad raw file.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\value{

}
\description{
Reads the configuration header from a Simrad raw0 file.
}
\examples{
\dontrun{}

}
