% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_GetDgHeader.R
\name{readEKRaw_GetDgHeader}
\alias{readEKRaw_GetDgHeader}
\title{Reads the a datagram header stored in a Simrad raw file.}
\usage{
readEKRaw_GetDgHeader(x, timeOffset = 0, endian = "little",
  xBase = -11644473600, offset = 0, ...)
}
\arguments{
\item{x}{The raw vector.}

\item{timeOffset}{is the time offset of the datagram.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31). NA}

\item{xBase}{is the base of Windows FILETIME: xBase=unclass(as.POSIXct('1601-1-1', tz="UTC"))[1].}

\item{offset}{an offset in bytes.}

\item{...}{Used for robustness.}
}
\value{

}
\description{
Reads the a datagram header stored in a Simrad raw file.
}
\examples{
\dontrun{}

}
