% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw_GetFileHeader}
\alias{readEKRaw_GetFileHeader}
\title{Read the header from a Simrad raw file.}
\usage{
readEKRaw_GetFileHeader(x, endian = "little", ...)
}
\arguments{
\item{x}{The raw vector.}

\item{endian}{The endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}

\item{...}{Used for robustness.}
}
\description{
Read the header from a Simrad raw file.
}
\examples{
\dontrun{}

}
