% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_GetRAW1_temp.R
\name{readEKRaw_GetRAW1}
\alias{readEKRaw_GetRAW1}
\title{Reads the sample data from a Simrad raw1 file.}
\usage{
readEKRaw_GetRAW1(x, endian = "little", complex.out = FALSE, ...)
}
\arguments{
\item{x}{is the path to the raw file.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\value{

}
\description{
Reads the sample data from a Simrad raw1 file.
}
\examples{
\dontrun{}

}
