% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw_GetSchema}
\alias{readEKRaw_GetSchema}
\title{Simrad raw schemas.}
\usage{
readEKRaw_GetSchema(
  dgName = c("RAW0", "RAW1", "ConfigHeader", "TransceiverConfig"),
  x = list(),
  var = "all",
  group = FALSE
)
}
\arguments{
\item{dgName}{The type of schema, such as a datagram name.}
}
\description{
Simrad raw schemas.
}
