% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw_GetSchemaOld}
\alias{readEKRaw_GetSchemaOld}
\title{Reads the sample data from a Simrad raw1 file.}
\usage{
readEKRaw_GetSchemaOld(
  dgName = c("RAW0", "RAW1", "ConfigHeader", "TransceiverConfig"),
  x = list(),
  var = "all"
)
}
\arguments{
\item{dgName}{The type of schema, such as a datagram name.}
}
\value{

}
\description{
Reads the sample data from a Simrad raw1 file.
}
\examples{
\dontrun{}

}
