% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw_power2sv.TSD}
\alias{readEKRaw_power2sv.TSD}
\title{Converts power to volume backscattering coefficient from data in the TSD format.}
\usage{
readEKRaw_power2sv.TSD(
  x,
  beams = list(),
  cali = NULL,
  list.out = FALSE,
  tiltcorr = 0,
  toTS = FALSE
)
}
\arguments{
\item{x}{is a list of the data (one element for each time step!!!).}

\item{beams}{is a list of the beam configuration of the sonar or echosounder.}

\item{cali}{optinal calibration information containing gain and Sa_correction values, possibly given pulse length values and frequencies.}

\item{list.out}{is TRUE to return the data as a list of acoustic values and calibration values.}

\item{tiltcorr}{is TRUE to apply the tilt correction used for fishery sonars.}

\item{toTS}{is TRUE to apply the TS calibration instead for the Sv calibration.}
}
\description{
Converts power to volume backscattering coefficient from data in the TSD format.
}
\examples{
\dontrun{}

}
