% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_ReadConfigHeader.R
\name{readEKRaw_ReadConfigHeader}
\alias{readEKRaw_ReadConfigHeader}
\title{Reads the configuration header from a Simrad raw0 file.}
\usage{
readEKRaw_ReadConfigHeader(fid, endian = "little")
}
\arguments{
\item{fid}{is the path to the raw file.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\value{

}
\description{
Reads the configuration header from a Simrad raw0 file.
}
\examples{
\dontrun{}

}
