% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_GetHeader.R, R/readEKRaw_ReadHeader.R
\name{readEKRaw_GetFileHeader}
\alias{readEKRaw_GetFileHeader}
\alias{readEKRaw_ReadHeader}
\title{Read the header from a Simrad raw file.}
\usage{
readEKRaw_GetFileHeader(x, endian = "little", ...)

readEKRaw_ReadHeader(fid, dgHeader, endian = "little")
}
\arguments{
\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}

\item{fid}{is the path to the raw file.}

\item{fid}{is the path to the raw file.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\value{



}
\description{
Read the header from a Simrad raw file.

Read the header from a Simrad raw file.
}
\examples{
\dontrun{}

\dontrun{}

}
