% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw_ReadSampledata_RAW0}
\alias{readEKRaw_ReadSampledata_RAW0}
\title{Reads the sample data from a Simrad raw0 file.}
\usage{
readEKRaw_ReadSampledata_RAW0(fid, endian = "little")
}
\arguments{
\item{fid}{is the path to the raw file.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\value{

}
\description{
Reads the sample data from a Simrad raw0 file.
}
\examples{
\dontrun{}

}
