% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_ReadTimes.R
\name{readEKRaw_ReadTimes}
\alias{readEKRaw_ReadTimes}
\title{Read all times of a raw file.}
\usage{
readEKRaw_ReadTimes(f, endian = "little", msg = TRUE, cores = 1)
}
\arguments{
\item{f}{A vector of raw file names or a directory containing raw files.}

\item{endian}{The endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}

\item{msg}{Logical: If TRUE print a time bar during reading.}

\item{cores}{An integer giving the number of cores to run in parallel.}
}
\value{

}
\description{
Read all times of a raw file.
}
\examples{
\dontrun{}

}
