% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw_ReadTransceiverConfig}
\alias{readEKRaw_ReadTransceiverConfig}
\title{Reads the transceiver configuration from a Simrad raw file.}
\usage{
readEKRaw_ReadTransceiverConfig(fid, endian = "little")
}
\arguments{
\item{fid}{is the path to the raw file.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\value{

}
\description{
Reads the transceiver configuration from a Simrad raw file.
}
\examples{
\dontrun{}

}
