% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_ScanDgHeaders.R
\name{readEKRaw_ScanDgHeaders}
\alias{readEKRaw_ScanDgHeaders}
\title{Read all datagram headers of a Simrad raw file.}
\usage{
readEKRaw_ScanDgHeaders(raw, endian = "little", msg = TRUE,
  rawsplit.out = FALSE)
}
\arguments{
\item{raw}{The path to directory of raw files or a vector of the paths to the raw files.}

\item{endian}{The endianness of the file, defaulted to "little".}

\item{msg}{Logical: if TRUE print a time bar during file reading.}

\item{rawsplit.out}{Logical: should the raw vector be split by datagrams and returned.}
}
\value{
A data frame of datagram info.
}
\description{
Provides the facility to modify parts of EK80/WBAT .raw files so that these files can be read by the Large Scale Survey System (LSSS) software.
}
