% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw_split}
\alias{readEKRaw_split}
\title{Splits Simrad raw files by beams.}
\usage{
readEKRaw_split(
  x,
  t = "all",
  beamnr = NULL,
  write = TRUE,
  newdir = NULL,
  nameadd = NULL,
  msg = FALSE,
  msgbar = TRUE,
  ext = "raw",
  ...
)
}
\arguments{
\item{x}{The path to a Simrad raw file or a directory of Simrad raw files.}

\item{t}{The time steps to extract the data for (use the default t="all" to keep all time steps)}

\item{beamnr}{A vector or list of beam indices. If empty (default) split into one file for each beam. If several beams should be in one file, use a list such as list(1, 2:4) for a four beam file, resulting in one file for the first beam and one file for the last three beams.}

\item{write}{Logical: If TRUE, write the splitted data into separate files (in separate directories as default).}

\item{newdir}{The path to the directory in which to put the splitted files, defaulted to one directory per beam in the same directory as the original raw files.}

\item{nameadd}{A string to be appended to the file names of the splitted files, defaulted to "Beams_1", "Beams_2,3,4" if splitting a four beam file into the first beam and the last three beams.}

\item{msg}{Logical: If TURE print a time bar for reading and writing of each individual raw file.}

\item{msgbar}{Logical: If TURE print a time bar showing the progression of splitting multiple files.}

\item{ext}{The file extension of the raw files.}

\item{...}{Further arguments passed on to list.files(), such as 'recursive'.}
}
\value{

}
\description{
Splits Simrad raw files by beams.
}
\examples{
\dontrun{}

}
