% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_stripNA.R
\name{readEKRaw_stripNA}
\alias{readEKRaw_stripNA}
\title{(Internal) Strips the input from NAs, used in EKRaw2TSD_oneFile().}
\usage{
readEKRaw_stripNA(x, NAind = NULL)
}
\value{

}
\description{
(Internal) Strips the input from NAs, used in EKRaw2TSD_oneFile().
}
\examples{
\dontrun{}

}
