% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRawOld.R
\name{readEKRawOld}
\alias{readEKRawOld}
\title{Reads a Simrad raw file.}
\usage{
readEKRawOld(f, t = 1, endian = "little", timeOffset = 0,
  minTimeDiff = Inf, drop.out = FALSE, msg = TRUE, prenumt = 10,
  na.rm = TRUE, cleartemp = FALSE)
}
\arguments{
\item{f}{is the path to the raw file.}

\item{t}{is a vector of the time steps to read.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31). NA}

\item{timeOffset}{is the time offset of the datagram.}

\item{minTimeDiff}{is the minimum difference in matlab time betwee two pings.}

\item{drop.out}{is TRUE to drop dimensions of the data.}

\item{msg}{is TRUE to print a time bar during reading.}

\item{prenumt}{is the number of time steps to reserve in the temporary file sto which data are saved during reading in order to restrict memeory usage. High values result in fewer temporary file but higher CPU time. The default value of 10 should be relatively optimal.}

\item{na.rm}{is TRUE to remove missing pings.}

\item{cleartemp}{is TRUE to clear all files in the temporary directory in which files are written in the TSD format for faster reading of the raw files. These files are deleted at startup of the operating system, but can be deleted specifically here if needed.}
}
\value{

}
\description{
Reads a Simrad raw file.
}
\examples{
\dontrun{}

}
