% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{SimradRaw}
\alias{SimradRaw}
\alias{SimradRaw-package}
\title{Read and write Simrad raw files}
\description{
This R package provides functions for reading, writing, splitting and other processing of Simrad raw files (raw0 (echosounder) and raw1 (fishery sonar)). The code is based on the Matlab library written by dr. Rick Towler, NOAA Alaska Fisheries Science Center. The functions use the TSD package for efficient reading of the data (writes to a temporary TSD file and reads afterwards to avoid appending in memory).
}
\details{
Details of the package
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/arnejohannesholmin/SimradRaw/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@imr.no}

}
