% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundbeam_range.R
\name{soundbeam_range}
\alias{soundbeam_range}
\title{This function calculates various variables related to the voxels of an aoustic system, such as the mid points or edge points of the voxels. NOTE: This function only considers one time step, which is a weekness, since the radial resolution may change between pings, particularly with fishery sonars.}
\usage{
soundbeam_range(x, pos = c("mid", "edge", "max", "res", "grid"),
  adds = NULL, Ro = NULL)
}
\arguments{
\item{x}{is a list of the variables 'lenb', and 'rres' or 'asps' and 'sint'.}

\item{pos}{is "res" to return the radial resolution, "max" to return the maximum range of the "mid" to return midpoints of the voxels and "edge" to return edge points, and numeric to return 'pos' positions between 0 and the maximum range to the edges.}

\item{adds}{A list of variables to add to the input list x.}

\item{Ro}{The range offset in meters.}
}
\value{

}
\description{
This function calculates various variables related to the voxels of an aoustic system, such as the mid points or edge points of the voxels. NOTE: This function only considers one time step, which is a weekness, since the radial resolution may change between pings, particularly with fishery sonars.
}
\examples{
\dontrun{}

}
