% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEKRaw.R, R/writeEKRaw_old.R
\name{writeEKRaw}
\alias{writeEKRaw}
\alias{writeEKRawOld}
\title{Writes a Simrad raw file.}
\usage{
writeEKRaw(data, con, header = NULL, t = 1, endian = "little",
  msg = TRUE)

writeEKRawOld(data, con, header = NULL, t = 1, endian = "little",
  msg = TRUE)
}
\arguments{
\item{data}{is a list of the data to write to the Simrad raw file.}

\item{con}{is the path to the raw file.}

\item{header}{is a list of the header to write to the Simrad raw file.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}

\item{msg}{is TRUE to print a time bar during reading.}

\item{t'}{is a vector of the time steps to write. NA}

\item{data}{is a list of the data to write to the Simrad raw file.}

\item{con}{is the path to the raw file.}

\item{header}{is a list of the header to write to the Simrad raw file.}

\item{t'}{is a vector of the time steps to write. NA}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}

\item{msg}{is TRUE to print a time bar during reading.}
}
\value{



}
\description{
Writes a Simrad raw file.

Writes a Simrad raw file.
}
\examples{
\dontrun{}

\dontrun{}

}
