% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEKRaw_WriteConfigHeader.R
\name{writeEKRaw_WriteConfigHeader}
\alias{writeEKRaw_WriteConfigHeader}
\title{Writes the configuration header to a Simrad raw0 file.}
\usage{
writeEKRaw_WriteConfigHeader(fid, configheader, endian = "little")
}
\arguments{
\item{fid}{is the path to the raw file.}

\item{configheader}{is a list of the configuration header.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\description{
Writes the configuration header to a Simrad raw0 file.
}
\examples{
\dontrun{}

}
