% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEKRaw_WriteDgHeader.R
\name{writeEKRaw_WriteDgHeader}
\alias{writeEKRaw_WriteDgHeader}
\title{Writes the sample data to a Simrad raw0 file.}
\usage{
writeEKRaw_WriteDgHeader(fid, dgType, dgTime, endian = "little", tz = "UTC")
}
\arguments{
\item{fid}{is the path to the raw file.}

\item{dgType}{is the datagram type.}

\item{dgTime}{is the time stamp of the datagram.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}

\item{tz}{is the the time zone. Overrides 'xBase' in mtim2FILETIME() if not given as tz="UTC". See as.POSIXlt().}
}
\description{
Writes the sample data to a Simrad raw0 file.
}
\examples{
\dontrun{}

}
