% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEKRaw_WriteHeader.R
\name{writeEKRaw_WriteHeader}
\alias{writeEKRaw_WriteHeader}
\title{Writes the header to a Simrad raw file.}
\usage{
writeEKRaw_WriteHeader(fid, config, dgLen, dgType, dgTime, endian = "little")
}
\arguments{
\item{fid}{is the path to the raw file.}

\item{config}{is a list of the transceiver configuration.}

\item{dgLen}{is the length of the datagram in bytes.}

\item{dgType}{is the datagram type.}

\item{dgTime}{is the time stamp of the datagram.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\description{
Writes the header to a Simrad raw file.
}
\examples{
\dontrun{}

}
