% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEKRaw_WriteSampledata_RAW0.R
\name{writeEKRaw_WriteSampledata_RAW0}
\alias{writeEKRaw_WriteSampledata_RAW0}
\title{Writes the sample data to a Simrad raw0 file.}
\usage{
writeEKRaw_WriteSampledata_RAW0(fid, data, endian = "little")
}
\arguments{
\item{fid}{is the path to the raw file.}

\item{data}{is a list of sample data to write to the Simrad raw file.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}
}
\value{

}
\description{
Writes the sample data to a Simrad raw0 file.
}
\examples{
\dontrun{}

}
