% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEKRaw_WriteTransceiverConfig.R
\name{writeEKRaw_WriteTransceiverConfig}
\alias{writeEKRaw_WriteTransceiverConfig}
\title{Writes the transceiver configuration to a Simrad raw file.}
\usage{
writeEKRaw_WriteTransceiverConfig(fid, configXcvr, i, endian = "little")
}
\arguments{
\item{fid}{is the path to the raw file.}

\item{configXcvr}{is a list of the transceiver configuration.}

\item{endian}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31).}

\item{xBase}{is the index of the time step.}
}
\value{

}
\description{
Writes the transceiver configuration to a Simrad raw file.
}
\examples{
\dontrun{}

}
