% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsetSamples.R
\name{offsetSamples}
\alias{offsetSamples}
\title{Shifts sonar data by an empirical range offset.}
\usage{
offsetSamples(x, beams)
}
\arguments{
\item{x}{The acoustic data to be shifted by the offset}

\item{beams}{A list containing the fields rofs (range offset in meters) and one of rres (range resolution) og the pair asps (average speed of sound) and sint (sampling interval).}
}
\value{

}
\description{
Shifts sonar data by an empirical range offset.
}
\examples{
\dontrun{}

}
