% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw}
\alias{readEKRaw}
\title{Reads a Simrad raw file.}
\usage{
readEKRaw(f, t = 1, endian = "little", timeOffset = 0, drop.out = FALSE,
  msg = TRUE, pb = TRUE, splitByPings = FALSE, complex.out = FALSE, ...)
}
\arguments{
\item{f}{The path to the raw file.}

\item{t}{A vector of the time steps to read.}

\item{timeOffset}{The time offset of the datagram.}

\item{drop.out}{Logical: If TRUE drop dimensions of the output data.}

\item{msg, pb}{See ?TSD::papply.}

\item{splitByPings}{Logical: If TRUE split the acoustic data into a list of single pings and strip off ranges with all NAs at the end of the beams (corresponding to the output from \code{\link{readEKRawOld}}).}

\item{complex.out}{Logical: If TRUE output also the complex values stored in the raw file for the RAW1 file format (fishery sonar).}

\item{...}{Used for robustness.}

\item{endian'}{The endian of the file, defaulted to "little".}
}
\value{

}
\description{
Reads a Simrad raw file.
}
\examples{
\dontrun{}

}
