% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw}
\alias{readEKRaw}
\title{Reads a Simrad raw file.}
\usage{
readEKRaw(f, t = 1, endian = "little", timeOffset = 0,
  drop.out = FALSE, msg = TRUE, splitByPings = FALSE,
  complex.out = FALSE, ...)
}
\arguments{
\item{f}{The path to the raw file.}

\item{t}{A vector of the time steps to read.}

\item{endian}{The endian of the file.}

\item{timeOffset}{The time offset of the datagram.}

\item{drop.out}{Logical: If TRUE drop dimensions of the data.}

\item{msg}{Logical: If TRUE print a time bar during reading.}

\item{splitByPings}{Logical: If TRUE split the acousic data by pings, which can avoid many NAs at the ends of the beams when the range shifts inside a file.}

\item{complex.out}{Logical: If TRUE output the real and imaginary part usd to extract power (for fishery sonars).}

\item{...}{Used for robustness.}
}
\value{

}
\description{
Reads a Simrad raw file.
}
\examples{
\dontrun{}

}
