% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw.R
\name{readEKRaw_extractBeams}
\alias{readEKRaw_extractBeams}
\title{Extracts one or more beams from raw data read by readEKRaw().}
\usage{
readEKRaw_extractBeams(x, beamnr = NULL, drop = FALSE)
}
\arguments{
\item{x}{A list of data read by readEKRaw().}

\item{beamnr}{A vector of indices of the beams to extract.}

\item{drop}{Logical: if TRUE drop the output when only one beam is extracted.}
}
\value{

}
\description{

}
\examples{
\dontrun{}

}
