% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_GetSchema.R
\name{readEKRaw_GetSchema}
\alias{readEKRaw_GetSchema}
\title{Reads the sample data from a Simrad raw1 file.}
\usage{
readEKRaw_GetSchema(dgName = c("RAW0", "RAW1", "ConfigHeader",
  "TransceiverConfig"))
}
\arguments{
\item{x}{is the path to the raw file.}
}
\value{

}
\description{
Reads the sample data from a Simrad raw1 file.
}
\examples{
\dontrun{}

}
