% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_GetDgHeader.R,
%   R/readEKRaw_ReadDgHeader.R
\name{readEKRaw_GetDgHeader}
\alias{readEKRaw_GetDgHeader}
\alias{readEKRaw_ReadDgHeader}
\title{Reads the a datagram header stored in a Simrad raw file.}
\usage{
readEKRaw_GetDgHeader(x, timeOffset = 0, endian = "little",
  xBase = -11644473600, offset = 0, ...)

readEKRaw_ReadDgHeader(fid, timeOffset = 0, endian = "little",
  xBase = -11644473600)
}
\arguments{
\item{timeOffset}{is the time offset of the datagram.}

\item{xBase}{is the base of Windows FILETIME: xBase=unclass(as.POSIXct('1601-1-1', tz="UTC"))[1].}

\item{fid}{is the path to the raw file.}

\item{endian'}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31). NA}

\item{fid}{is the path to the raw file.}

\item{timeOffset}{is the time offset of the datagram.}

\item{endian'}{is the endian of the file, defaulted to .Platform$endian (changed from "big" by Arne Johannes Holmin 2012-07-31). NA}

\item{xBase}{is the base of Windows FILETIME: xBase=unclass(as.POSIXct('1601-1-1', tz="UTC"))[1].}
}
\value{



}
\description{
Reads the a datagram header stored in a Simrad raw file.

Reads the a datagram header stored in a Simrad raw file.
}
\examples{
\dontrun{}

\dontrun{}

}
