% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_ScanDgHeaders.R
\name{readEKRaw_ScanDgHeaders}
\alias{readEKRaw_ScanDgHeaders}
\title{Read all datagram headers of a Simrad raw file.}
\usage{
readEKRaw_ScanDgHeaders(filename, endian = "little", msg = TRUE)
}
\arguments{
\item{filename}{The path to a raw file.}

\item{endian}{the endianness of the file, defaulted to "little".}

\item{msg}{logical: if TRUE print a time bar during file reading.}
}
\value{
A data frame of datagram info.
}
\description{
Read all datagram headers of a Simrad raw file.
}
