% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEKRaw_ScanDgHeaders.R,
%   R/readEKRaw_ScanDgHeaders_old.R
\name{readEKRaw_ScanDgHeaders}
\alias{readEKRaw_ScanDgHeaders}
\alias{readEKRaw_ScanDgHeaders_old}
\title{Modify parts of EK80/WBAT .raw files}
\usage{
readEKRaw_ScanDgHeaders(raw, endian = "little", msg = TRUE,
  raw.out = FALSE)

readEKRaw_ScanDgHeaders_old(filename, t = "all", addMRU = TRUE,
  fixBeamType = TRUE, fixSlope = TRUE, endian = "little", msg = TRUE)
}
\arguments{
\item{endian}{the endianness of the file, defaulted to "little".}

\item{msg}{logical: if TRUE print a time bar during file conversion.}

\item{filename}{The path to directory of raw files or a vector of the paths to the raw files.}

\item{addMRU}{logical; if TRUE the funciton inserts a dummy MRU0 datagram prior to RAW3 datagrams.}

\item{fixBeamType}{logical; if TRUE the BeamType is fixed to 1.}

\item{fixSlope}{logical; if TRUE the slope is set to a default.}

\item{addTransducerSeralNumber}{A string giving the transducer serial number.}

\item{filename}{The path to directory of raw files or a vector of the paths to the raw files.}

\item{addTransducerSeralNumber}{A string giving the transducer serial number.}

\item{endian}{the endianness of the file, defaulted to "little".}

\item{msg}{logical: if TRUE print a time bar during file conversion.}
}
\value{
A vector of file names of the modified files.

A vector of file names of the modified files.
}
\description{
Provides the facility to modify parts of EK80/WBAT .raw files so that these files can be read by the Large Scale Survey System (LSSS) software.

Provides the facility to modify parts of EK80/WBAT .raw files so that these files can be read by the Large Scale Survey System (LSSS) software.
}
